% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Agregar Resultados Nacionales desede Simulaciones Provinciales}
\usage{
aggregate_results(votos_provincias_sims, interval_level = 0.9)
}
\arguments{
\item{votos_provincias_sims}{Data frame con resultados provinciales por simulacion}

\item{interval_level}{Nivel de confianza para los intervalos (default 0.9)}
}
\value{
Data frame con resultados nacionales agregados
}
\description{
Agrega los resultados de votos y esca\enc{ñ}{n}os a nivel nacional
}
