% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrapping de datos de encuesta}
\usage{
bootstrap(data, B, calib_vars, weights)
}
\arguments{
\item{data}{Data frame con datos de encuesta (columnas: recuerdo, idv, ponde)}

\item{B}{Numero de replicas a generar}

\item{calib_vars}{Variables para calibracion (vector de nombres de columnas)}

\item{weights}{Nombre de la columna con pesos de referencia para calibracion}
}
\value{
Data frame con replicas de datos de encuesta
}
\description{
Genera replicas de datos de encuesta usando bootstrapping con reemplazo
}
\keyword{internal}
