% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_n}
\alias{calculate_n}
\title{Calcular Tama\enc{ñ}{n}os de Muestra de MT}
\usage{
calculate_n(mt)
}
\arguments{
\item{mt}{Data frame con matriz de transferencia}
}
\value{
Data frame con fila N a\enc{ñ}{n}adida
}
\description{
Calcula los tama\enc{ñ}{n}os de muestra efectivos para cada columna de recuerdo
en la matriz de transferencia.
}
