% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{check_factor_lvls}
\alias{check_factor_lvls}
\title{Validar Niveles de Factores}
\usage{
check_factor_lvls(
  recuerdo_var = NULL,
  recuerdo_lvls = NULL,
  idv_var = NULL,
  idv_lvls = NULL
)
}
\arguments{
\item{recuerdo_var}{Vector con valores observados de recuerdo (opcional)}

\item{recuerdo_lvls}{Vector con niveles validos de recuerdo (opcional)}

\item{idv_var}{Vector con valores observados de IDV (opcional)}

\item{idv_lvls}{Vector con niveles validos de IDV (opcional)}
}
\value{
Invisible(TRUE) si es valida, o error si hay valores no previstos
}
\description{
Valida que los valores observados en variables de recuerdo e IDV
esten dentro de las categorias esperadas.
}
\keyword{internal}
