% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{check_mt}
\alias{check_mt}
\title{Validate Vote Transfer Matrix}
\usage{
check_mt(mt, recuerdo_lvls, idv_lvls)
}
\arguments{
\item{mt}{Data frame with transfer matrix}

\item{recuerdo_lvls}{Vector with valid recuerdo codes}

\item{idv_lvls}{Vector with valid IDV codes}
}
\value{
Invisible(TRUE) if valid, or error if not
}
\description{
Validates that party codes in the transfer matrix
correspond to expected categories.
}
\keyword{internal}
