% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{check_nseats}
\alias{check_nseats}
\title{Validar Correspondencia de escanos por Provincia}
\usage{
check_nseats(n_seats, patrones, strategy = "top_down")
}
\arguments{
\item{n_seats}{Data frame con escanos por provincia}

\item{patrones}{Data frame con patrones provinciales}
}
\value{
Invisible(TRUE) si es valida, warning/error segun corresponda
}
\description{
Valida que las provincias en los datos de escanos correspondan
con las provincias en los patrones provinciales.
}
\keyword{internal}
