% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{check_patrones}
\alias{check_patrones}
\title{Validate District Patterns}
\usage{
check_patrones(patrones, idv_lvls)
}
\arguments{
\item{patrones}{Data frame with district patterns}

\item{idv_lvls}{Vector with valid IDV codes}
}
\value{
Invisible(TRUE) if valid, or error if not
}
\description{
Validates that IDV names in district patterns are valid.
}
\keyword{internal}
