% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{check_retoques}
\alias{check_retoques}
\title{Validar Retoques (Ajustes Manuales)}
\usage{
check_retoques(retoques, idv_lvls)
}
\arguments{
\item{retoques}{Data frame con ajustes manuales}

\item{idv_lvls}{Vector con codigos IDV validos}
}
\value{
Invisible(TRUE) si es valida, o error si no
}
\description{
Valida que los codigos de partido en retoques esten
dentro de los codigos IDV validos.
}
\keyword{internal}
