% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{check_small_parties}
\alias{check_small_parties}
\title{Validar Partidos pequenos}
\usage{
check_small_parties(small_parties, idv_lvls)
}
\arguments{
\item{small_parties}{Data frame con partidos pequenos}

\item{idv_lvls}{Vector con codigos IDV validos}
}
\value{
Invisible(TRUE) si es valida, o error si no
}
\description{
Valida que los codigos de partido en small_parties esten
dentro de los codigos IDV validos.
}
\keyword{internal}
