% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{check_votos_ant}
\alias{check_votos_ant}
\title{Validate Previous Election Data}
\usage{
check_votos_ant(votos_ant, recuerdo_lvls)
}
\arguments{
\item{votos_ant}{Data frame with previous electoral results}

\item{recuerdo_lvls}{Vector with valid 'recuerdo' codes}
}
\value{
Invisible(TRUE) si es valida, o error si no
}
\description{
Validates that previous election data contains
all necessary party codes.
}
\keyword{internal}
