% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{create_input_template}
\alias{create_input_template}
\title{Crear plantilla de archivos de entrada}
\usage{
create_input_template(file_path = "input/input.xlsx")
}
\arguments{
\item{file_path}{Path of Excel file to create}
}
\value{
Returns (invisibly) path to created file
}
\description{
Genera un archivo Excel con todas las hojas necesarias para el analisis electoral.
}
\details{
Sheets created and minimum columns required by package:
\itemize{
\item partidos: idv, recuerdo
\item mt_simplificada: idv + recuerdo columns + 'N' row
\item patrones: codigo_provincia + party columns (percentage 0-100)
\item anteriores_elecciones: recuerdo, votos_ant
\item n_diputados: codigo_provincia, n_diputados
\item retoques: idv, votos_adicionales
\item small_parties: idv, votos
}
}
\seealso{
Other utility-functions: 
\code{\link{setup_electoral_project}()}

Other project-setup: 
\code{\link{setup_electoral_project}()}
}
\concept{project-setup}
\concept{utility-functions}
