% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_simulation_pipeline.R
\name{execute_vota}
\alias{execute_vota}
\title{Execute VOTA algorithm based on the selected strategy}
\usage{
execute_vota(strategy = c("top_down", "bottom_up"), ...)
}
\arguments{
\item{strategy}{Strategy for simulation ("top_down" or "bottom_up").}

\item{...}{Additional arguments to be passed to the internal functions.}
}
\value{
A list containing the results of the VOTA algorithm execution.
}
\description{
This function executes the VOTA algorithm using either a top-down or bottom-up approach,
depending on the specified strategy. It captures and validates additional arguments
and calls the appropriate internal function to perform the simulation.
}
\keyword{internal}
