% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_censo}
\alias{get_censo}
\title{Download Electoral Census Data}
\usage{
get_censo(provincias)
}
\arguments{
\item{provincias}{Vector with province codes to obtain census data for}
}
\value{
Data frame with columns codigo_provincia and censo_real
}
\description{
Downloads electoral census data from INE for specified provinces.
}
\details{
This function automatically downloads the census file from INE and extracts
data for requested provinces. Requires internet connection.
}
\examples{
\dontrun{
# Get census data for Madrid, Barcelona and Valencia
censo <- get_censo(c("28", "08", "46"))
}

}
