% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_replica_df}
\alias{get_replica_df}
\title{Extrae una replica de un objeto svrepdesign}
\usage{
get_replica_df(des_rep, j)
}
\arguments{
\item{des_rep}{Objeto svrepdesign con replicas}

\item{j}{Numero de replica a extraer}
}
\value{
Data frame con datos de la replica j
}
\description{
Extrae una replica de un objeto svrepdesign
}
\keyword{internal}
