% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_cer_url}
\alias{last_cer_url}
\title{Get Latest Electoral Census URL}
\usage{
last_cer_url()
}
\value{
A string containing the URL of the latest electoral census file.
}
\description{
Fetches the URL of the latest electoral census file from the INE website. If there is an error during fetching, it returns a fallback URL.
}
\keyword{internal}
