% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{load_electoral_data_bottom_up}
\alias{load_electoral_data_bottom_up}
\title{Cargar Datos Electorales Bottom-Down desde Excel}
\usage{
load_electoral_data_bottom_up(input_path, uncertainty_method)
}
\arguments{
\item{input_path}{Ruta al archivo Excel de entrada}

\item{uncertainty_method}{Solo puede ser "bootstrap"}
}
\value{
Lista con todos los datos de entrada validados
}
\description{
Carga todos los datos necesarios desde un archivo Excel con multiples hojas
}
\details{
Hojas requeridas en el Excel:
\itemize{
\item anteriores_elecciones: resultados electorales anteriores
\item n_diputados: escanos por provincia
\item retoques: ajustes manuales (opcional)
\item small_parties: partidos pequenos (opcional)
}
}
\keyword{internal}
