% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt}
\alias{mt}
\title{Matriz de Transferencia de Ejemplo}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{idv}{Factor con la intención de voto de los partidos}
\item{PSOE}{Numérico, porcentaje de transferencia desde recuerdo PSOE}
\item{PP}{Numérico, porcentaje de transferencia desde recuerdo PP}
\item{Vox}{Numérico, porcentaje de transferencia desde recuerdo Vox}
\item{Sumar}{Numérico, porcentaje de transferencia desde recuerdo Sumar}
\item{OTBL}{Numérico, porcentaje de transferencia desde recuerdo OTBL}
\item{ABNL}{Numérico, porcentaje de transferencia desde recuerdo ABNL}
}
}
\source{
Datos sintéticos de demostración
}
\usage{
mt
}
\description{
Dataset de ejemplo con una matriz de transferencia electoral empleada
en las plantillas del paquete.
}
\seealso{
Other example-data: 
\code{\link{n_seats}},
\code{\link{patrones_23J}},
\code{\link{retoques}},
\code{\link{small_parties}},
\code{\link{votos_23J}}
}
\concept{example-data}
\keyword{datasets}
