% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mt_correction}
\alias{mt_correction}
\title{Apply Corrections to Vote Transfer Matrix}
\usage{
mt_correction(
  mt_simplificada,
  factor_correccion_otbl,
  factor_correccion_abstencion,
  factor_correccion_jovenes
)
}
\arguments{
\item{mt_simplificada}{Data frame with vote transfer matrix}

\item{factor_correccion_otbl}{Correction factor for other blank/null votes}

\item{factor_correccion_abstencion}{Correction factor for abstention}

\item{factor_correccion_jovenes}{Correction factor for new voters}
}
\value{
Data frame with corrected transfer matrix
}
\description{
Auxiliary function that applies corrections for abstention, new voters
and other blank/null votes to the transfer matrix.
}
\keyword{internal}
