% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_seats}
\alias{n_seats}
\title{Número de escaños por provincia}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{codigo_provincia}{Carácter, código INE de la provincia}
\item{n_diputados}{Entero, número de escaños asignados}
}
}
\source{
Instituto Nacional de Estadística (INE)
}
\usage{
n_seats
}
\description{
Dataset con el número de diputados asignados a cada provincia española.
}
\seealso{
Other example-data: 
\code{\link{mt}},
\code{\link{patrones_23J}},
\code{\link{retoques}},
\code{\link{small_parties}},
\code{\link{votos_23J}}
}
\concept{example-data}
\keyword{datasets}
