% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-electo_fit.R
\name{new_electo_fit}
\alias{new_electo_fit}
\title{Crea un objeto electo_fit}
\usage{
new_electo_fit(results)
}
\arguments{
\item{results}{Lista con los resultados del ajuste, que debe incluir los siguientes elementos:
\describe{
\item{estimacion}{Data frame con las estimaciones finales, debe incluir columnas 'partido', 'pct_m', 'seats_m'}
\item{estimacion_sims}{(Opcional) Data frame con las simulaciones detalladas por partido y provincia}
\item{estimacion_provincias_sims}{(Opcional) Data frame con las simulaciones detalladas por partido y provincia}
\item{mt_sims_pct}{(Opcional) Matriz de transferencias simuladas en porcentaje de voto sobre recuerdo}
\item{mt_sims_electores}{(Opcional) Matriz
de transferencias simuladas en número de electores}
\item{dhondt_output}{(Opcional) Resultado del método D'Hondt aplicado a las simulaciones}
\item{participacion_media}{(Opcional) Valor numérico con la participación media estimada}
\item{metadata}{Lista con metadatos del ajuste, como 'nsims', 'tau', 'umbral', etc.}
}}
}
\value{
Objeto de clase electo_fit
}
\description{
Crea un objeto electo_fit
}
