% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patrones_23J}
\alias{patrones_23J}
\title{Patrones electorales del 23J}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{codigo_provincia}{Carácter, código INE de la provincia}
\item{partido}{Carácter, código del partido político}
\item{patron}{Numérico, proporción histórica de votos (0-1)}
}
}
\source{
Ministerio del Interior, elecciones generales 23J 2023
}
\usage{
patrones_23J
}
\description{
Dataset con patrones históricos de voto por provincia para las elecciones
generales del 23 de julio de 2023.
}
\seealso{
Other example-data: 
\code{\link{mt}},
\code{\link{n_seats}},
\code{\link{retoques}},
\code{\link{small_parties}},
\code{\link{votos_23J}}
}
\concept{example-data}
\keyword{datasets}
