% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Forward-pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{Left-hand side value}

\item{rhs}{Right-hand side expression}
}
\description{
Re-exports magrittr's pipe so '\%>\%' is available to package users
}
\keyword{internal}
