% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-electo_fit.R
\name{plot.electo_fit}
\alias{plot.electo_fit}
\title{Generate plots from electo_fit objects}
\usage{
\method{plot}{electo_fit}(
  x,
  kind = c("nacional", "seats_dist", "provincia", "dhondt_margin"),
  partido = NULL,
  ...
)
}
\arguments{
\item{x}{Objeto electo_fit}

\item{kind}{"nacional", "seats_dist", "provincia", "dhondt_margin"}

\item{partido}{Party to highlight (used when kind="provincia")}

\item{...}{Additional arguments (not used)}
}
\description{
Generate plots from electo_fit objects
}
\concept{electo-fit-methods}
\concept{visualization-functions}
