% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_pipeline_args}
\alias{prepare_pipeline_args}
\title{Prepare Pipeline Arguments}
\usage{
prepare_pipeline_args(params, ...)
}
\arguments{
\item{params}{A list containing various parameter settings for the simulation.}

\item{...}{Additional arguments to be included in the output list.}
}
\value{
A list of arguments ready to be passed to the simulation pipeline.
}
\description{
Prepares a list of arguments for the election simulation pipeline based on the provided parameters.
}
\keyword{internal}
