% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-electo_fit.R
\name{print.electo_fit}
\alias{print.electo_fit}
\title{Print method for electo_fit objects}
\usage{
\method{print}{electo_fit}(x, ...)
}
\arguments{
\item{x}{Objeto electo_fit}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for electo_fit objects
}
