% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-electo_fit.R
\name{print.summary_electo_fit}
\alias{print.summary_electo_fit}
\title{Print method for summary_electo_fit objects}
\usage{
\method{print}{summary_electo_fit}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object summary_electo_fit}

\item{digits}{Number of digits to print}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for summary_electo_fit objects
}
