% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{read_partidos}
\alias{read_partidos}
\title{Cargar codigos de Partidos}
\usage{
read_partidos(input_path)
}
\arguments{
\item{input_path}{Ruta al archivo Excel}
}
\value{
Lista con vectores 'recuerdo' e 'idv' de codigos de partidos
}
\description{
Extrae los codigos de partidos desde la hoja 'partidos' del Excel.
}
