% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retoques}
\alias{retoques}
\title{Ajustes manuales de ejemplo}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{idv}{Carácter, código del partido}
\item{votos_adicionales}{Entero, votos a añadir (positivo) o restar (negativo)}
}
}
\source{
Datos sintéticos de demostración
}
\usage{
retoques
}
\description{
Dataset con retoques o ajustes manuales para aplicar a las estimaciones
electorales.
}
\seealso{
Other example-data: 
\code{\link{mt}},
\code{\link{n_seats}},
\code{\link{patrones_23J}},
\code{\link{small_parties}},
\code{\link{votos_23J}}
}
\concept{example-data}
\keyword{datasets}
