% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{setup_electoral_project}
\alias{setup_electoral_project}
\title{Set up electoral project}
\usage{
setup_electoral_project(project_dir)
}
\arguments{
\item{project_dir}{Project directory (will be created if it doesn't exist)}
}
\value{
Invisible(NULL). Function is executed for its side effects.
}
\description{
Creates directory structure and template files for a new electoral project.
}
\details{
This function:
\itemize{
\item Creates input/, output/ and scripts/ folders
\item Generates an Excel file with all sheets required by the pipeline
\item Creates a main example script in scripts/main.R
}
}
\examples{
\dontrun{
setup_electoral_project("my_simulation_2023")
}

}
\seealso{
Other utility-functions: 
\code{\link{create_input_template}()}

Other project-setup: 
\code{\link{create_input_template}()}
}
\concept{project-setup}
\concept{utility-functions}
