% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_parties}
\alias{small_parties}
\title{Partidos pequeños de ejemplo}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{idv}{Carácter, código del partido}
\item{votos}{Entero, estimación de votos a nivel nacional}
}
}
\source{
Datos sintéticos de demostración
}
\usage{
small_parties
}
\description{
Dataset con estimaciones de voto para partidos pequeños no incluidos en la
matriz de transferencia principal.
}
\seealso{
Other example-data: 
\code{\link{mt}},
\code{\link{n_seats}},
\code{\link{patrones_23J}},
\code{\link{retoques}},
\code{\link{votos_23J}}
}
\concept{example-data}
\keyword{datasets}
