% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-electo_fit.R
\name{summary.electo_fit}
\alias{summary.electo_fit}
\title{Summary method for electo_fit objects}
\usage{
\method{summary}{electo_fit}(object, ...)
}
\arguments{
\item{object}{Object electo_fit}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for electo_fit objects
}
