% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbral_electoral.R
\name{umbral_electoral}
\alias{umbral_electoral}
\title{Aplica el umbral electoral a los datos de votos simulados por provincias.}
\usage{
umbral_electoral(
  votos_provincias_sims,
  umbral = 0.03,
  tipo_umbral = "provincial"
)
}
\arguments{
\item{votos_provincias_sims}{Data frame con votos simulados por provincia}

\item{umbral}{Umbral minimo de voto para asignacion de escaños (por defecto 0.03)}

\item{tipo_umbral}{Tipo de umbral: "provincial", "autonomico" o "mixto" (por defecto "provincial")}
}
\value{
Lista con dos data frames:
\describe{
\item{noentran}{Partidos que no alcanzan el umbral, con votos y seats = NA}
\item{entran}{Partidos que superan el umbral, con votos para asignacion D'Hondt}
}
}
\description{
Aplica el umbral electoral a los datos de votos simulados por provincias.
}
\keyword{internal}
