% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-electo_fit.R
\name{validate_electo_fit}
\alias{validate_electo_fit}
\title{Valida un objeto electo_fit}
\usage{
validate_electo_fit(x)
}
\arguments{
\item{x}{Objeto a validar}
}
\value{
TRUE si es válido, error si no lo es
}
\description{
Valida un objeto electo_fit
}
\keyword{internal}
