% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{validate_params}
\alias{validate_params}
\title{Validate Parameters}
\usage{
validate_params(
  survey_data = NULL,
  input_path,
  output_file = NULL,
  uncertainty_method = c("mcmc", "bootstrap"),
  strategy = c("top_down", "bottom_up"),
  nsims = 100,
  factor_correccion_abstencion = 3,
  factor_correccion_jovenes = 2.5,
  factor_correccion_otbl = 3,
  tiempo_entre_elecciones = 0.1,
  district_col = "codigo_provincia",
  tau = 300,
  umbral = 0.03,
  tipo_umbral = "provincial",
  interval_level = 0.9,
  censo = NULL,
  verbose = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{survey_data}{Data frame with individual survey data (required if uncertainty_method is "bootstrap")}

\item{input_path}{Path to Excel input file with multiple sheets}

\item{output_file}{Path to the output RDS file where results will be saved (e.g., "output/results.rds")}

\item{uncertainty_method}{Type of input data: "bootstrap" or "mcmc"}

\item{strategy}{Projection strategy: "top_down" or "bottom_up"}

\item{nsims}{Number of Monte Carlo simulations for transfer matrices (or bootstrap replications if uncertainty_method is "bootstrap")}

\item{factor_correccion_abstencion}{Abstention correction factor (default 3)}

\item{factor_correccion_jovenes}{New voters correction factor (default 2.5)}

\item{factor_correccion_otbl}{Other+blank votes intention correction factor (default 3)}

\item{tiempo_entre_elecciones}{Years between elections for demographic adjustments}

\item{district_col}{Name of column in survey_data indicating province or electoral district}

\item{tau}{Parameter to control variability level of provincial party projection patterns (only needed if strategy is "top_down")}

\item{umbral}{Minimum vote threshold for seat assignment (default 0.03)}

\item{tipo_umbral}{Threshold type: "provincial", "autonomico" or "mixto" (default "provincial")}

\item{interval_level}{Confidence level for uncertainty intervals (default 0.9)}

\item{verbose}{Show progress messages (default TRUE)}

\item{seed}{Seed for reproducibility (default NULL)}

\item{...}{Additional arguments for bootstrap internal function. The current accepted parameters are calib_vars and weights. Only used when uncertainty_method is "bootstrap"}
}
\value{
List with organized and validated parameters for the simulation pipeline
}
\description{
Validates and organizes input parameters for the electoral simulation pipeline.
}
\keyword{internal}
