% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{votos_23J}
\alias{votos_23J}
\title{Resultados electorales del 23J}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{recuerdo}{Carácter, código del partido según recuerdo de voto}
\item{votos_ant}{Entero, votos obtenidos en la elección anterior}
}
}
\source{
Ministerio del Interior, elecciones generales 23J 2023
}
\usage{
votos_23J
}
\description{
Dataset con los resultados oficiales de las elecciones generales del 23 de
julio de 2023, utilizados como referencia de elecciones anteriores.
}
\seealso{
Other example-data: 
\code{\link{mt}},
\code{\link{n_seats}},
\code{\link{patrones_23J}},
\code{\link{retoques}},
\code{\link{small_parties}}
}
\concept{example-data}
\keyword{datasets}
